requirements.txt
lxml
bs4
tqdm
slackclient
roshi
requests
